+++ src/main/java/com/smile/petpat/image/domain/ImageUploadManager.java
@@ -2,10 +2,12 @@ package com.smile.petpat.image.domain;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.smile.petpat.image.repository.ImageRepository;
+import com.smile.petpat.post.category.domain.PostType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -23,14 +25,14 @@ public class ImageUploadManager {
 
     }
 
-    /*게시글 이미지 등록 */
-    public List<String> uploadPostImage(List<MultipartFile> multipartFiles){
+    /* 게시글 이미지 등록 */
+    public List<String> uploadPostImage(List<MultipartFile> multipartFiles,Long postId, PostType postType ){
         List<Image> imageList = new ArrayList<>();
         multipartFiles.forEach(file ->{
             String fakeFileName = imageUtils.createFileName(file.getOriginalFilename());
             String originalFileName = file.getOriginalFilename();
             String filePath = s3Uploader.uploadFile(file);
-            Image image = imageUploader.toImageEntity(fakeFileName,originalFileName,filePath);
+            Image image = imageUploader.toImageEntity(fakeFileName,originalFileName,filePath,postId,postType);
             imageList.add(image);
         });
 
@@ -39,7 +41,25 @@ public class ImageUploadManager {
                 .collect(Collectors.toCollection(ArrayList :: new));
     }
 
+    // 이미지 파일 수정
+    @Transactional
+    public List<String> updateFile(List<MultipartFile> multipartFiles, Long postId, PostType postType) {
+        List<String> fakeFiles = imageUploader.createKey(postId, postType);
+        for (String fakeFile :fakeFiles) {
+            imageUtils.getFileExtension(fakeFile);
+        }
+        s3Uploader.deleteS3(fakeFiles,postId, postType);
+        imageUploader.deleteImg(postId, postType);
+        return uploadPostImage(multipartFiles, postId, postType);
+    }
 
+    // 게시글 삭제 이미지도 삭제
+    public void removePostImage( Long postId, PostType postType){
+        List<String> fakeFiles = imageUploader.createKey(postId, postType);
+        s3Uploader.deleteS3(fakeFiles,postId, postType);
+        imageUploader.deleteImg(postId, postType);
+
+    }
