+++ src/main/java/com/smile/petpat/post/qna/service/QnaServiceImpl.java
@@ -1,4 +1,7 @@
 package com.smile.petpat.post.qna.service;
 
+import org.springframework.stereotype.Service;
+
+@Service
 public class QnaServiceImpl {
 }
